# docs/designs/api/US-001-openapi.yaml
openapi: 3.0.3
info:
  title: Customer Portal - Customer Registration API
  description: Public API endpoint for customer self-registration in the Customer Portal application.
  version: "1"
x-story: US-001
x-source-specification: docs/specifications/US-001-spec.md
x-source-specification-version: 1
x-produced-by: openapi-designer

servers:
  - url: /api/v1
    description: Current API version root

paths:
  /customers:
    post:
      summary: Register a new customer
      description: |
        Creates a new customer account using an email address and password.
        The endpoint is public and does not require prior authentication.
        On success, creates a customer resource, assigns the CUSTOMER role,
        and returns a 201 Created status with the Location header.
      operationId: registerCustomer
      tags:
        - Customer Registration
      security: [] # Public endpoint - unauthenticated per SC-4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterCustomerRequest'
      responses:
        '201':
          description: Customer account successfully created.
          headers:
            Location:
              description: Relative URI path of the created customer resource.
              schema:
                type: string
                example: /api/v1/customers/1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Validation error or malformed request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '409':
          description: Account with the specified email already exists (duplicate conflict).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-08-31T15:54:00Z"
                status: 409
                error: Conflict
                message: "An account with this email already exists."
                path: /api/v1/customers
        '415':
          description: Unsupported Media Type (Content-Type header is not application/json).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-08-31T15:54:00Z"
                status: 415
                error: Unsupported Media Type
                message: "Content-Type must be application/json"
                path: /api/v1/customers
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-08-31T15:54:00Z"
                status: 500
                error: Internal Server Error
                message: "An unexpected error occurred"
                path: /api/v1/customers

components:
  schemas:
    RegisterCustomerRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: Valid and unique email address for the customer account.
          example: customer@example.com
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 72
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[!@#$%^&*()_+\-=\[\]{};'':"\\|,.<>\/?]).{12,72}$'
          description: Plaintext password meeting complexity requirements (at least 1 uppercase, 1 lowercase, 1 digit, 1 special character).
          example: "SecurePass123!"

    CustomerResponse:
      type: object
      required:
        - id
        - email
        - role
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: Unique surrogate identifier of the created customer.
          example: 1
        email:
          type: string
          format: email
          description: Normalized email address of the customer.
          example: customer@example.com
        role:
          type: string
          enum:
            - CUSTOMER
          description: Default role assigned to the customer.
          example: CUSTOMER
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp in UTC ISO-8601.
          example: "2026-08-31T15:54:00Z"

    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          description: Name of the invalid request field.
          example: email
        message:
          type: string
          description: Validation error message for the field.
          example: Invalid email format

    ValidationErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-08-31T15:54:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: "Validation failed"
        path:
          type: string
          example: /api/v1/customers
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'

    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-08-31T15:54:00Z"
        status:
          type: integer
          example: 409
        error:
          type: string
          example: Conflict
        message:
          type: string
          example: "An account with this email already exists."
        path:
          type: string
          example: /api/v1/customers