# src/main/java/org/example/customerportal/model/dto/CustomerResponse.java
package org.example.customerportal.model.dto;

import lombok.*;

import java.time.Instant;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class CustomerResponse {

    private Long id;
    private String email;
    private String role;
    private Instant createdAt;
}